+++ src/control_plane/migration/v7/v7_logical_data_migration_source_options.rs
@@ -8,5 +8,6 @@ pub(crate) struct V7LogicalDataMigrationSourceOptions {
     pub(crate) driver: String,
     pub(crate) container_name: String,
     pub(crate) container_id: String,
+    pub(crate) named_volumes: Vec<String>,
     pub(crate) logical_data: BTreeMap<String, String>,
 }